agrox_backend/requirements.txt
fastapi
uvicorn[standard]
python-multipart
pillow
numpy
tensorflow
opencv-python
scikit-learn
joblib
scipy
requests
python-dotenv
openai
firebase-admin
pydantic